feat: FileHelper Cleanup
The remaining fileHelper files have been removed. The remaining tests for the
expand_absolute_path function have been moved. A call to configure_target has
been added in CMakeLists.txt file, which required some changes to avoid warnings
and compilation errors.
