+++ src/common/data_provider/CMakeLists.txt
@@ -94,7 +94,6 @@ include_directories(${SRC_FOLDER}/common/utils/)
 include_directories(${SRC_FOLDER}/common/byteArrayHelper/include/)
 include_directories(${SRC_FOLDER}/common/cmdHelper/include/)
 include_directories(${SRC_FOLDER}/common/globHelper/include/)
-include_directories(${SRC_FOLDER}/common/fileHelper/include/)
 include_directories(${SRC_FOLDER}/common/jsonHelper/include/)
 include_directories(${SRC_FOLDER}/common/linuxHelper/include/)
 include_directories(${SRC_FOLDER}/common/stringHelper/include/)
