--- src/common/fileHelper/CMakeLists.txt
@@ -1,5 +1,0 @@
-cmake_minimum_required(VERSION 3.12.4)
-
-if(UNIT_TEST)
-    add_subdirectory(tests)
-endif()
