+++ src/common/file_handler/CMakeLists.txt
@@ -3,6 +3,7 @@ cmake_minimum_required(VERSION 3.22)
 project(FileHandler)
 
 include(../../cmake/CommonSettings.cmake)
+include(../../cmake/ConfigureTarget.cmake)
 set_common_settings()
 
 # Filesystem target
@@ -11,8 +12,10 @@ add_library(Filesystem STATIC
     ${CMAKE_CURRENT_SOURCE_DIR}/filesystem/src/filesystem.cpp
 )
 
-target_include_directories(Filesystem PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/filesystem/include)
+target_include_directories(Filesystem PUBLIC
+    ${CMAKE_CURRENT_SOURCE_DIR}/filesystem/include PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/filesystem/src)
 target_link_libraries(Filesystem PUBLIC utils)
+configure_target(Filesystem)
 
 # File io target
 
@@ -20,5 +23,11 @@ add_library(FileIO STATIC
     ${CMAKE_CURRENT_SOURCE_DIR}/file_io/src/file_io.cpp
 )
 
-target_include_directories(FileIO PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/file_io/include)
+target_include_directories(FileIO PUBLIC
+    ${CMAKE_CURRENT_SOURCE_DIR}/file_io/include PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/file_io/src)
+configure_target(FileIO)
 
+if(BUILD_TESTS)
+    enable_testing()
+    add_subdirectory(filesystem/tests)
+endif()
