+++ src/common/file_handler/file_io/src/file_io.cpp
@@ -41,27 +41,32 @@ namespace file_io
 
     std::vector<char> FileIO::getBinaryContent(const std::string& filePath) const
     {
-        auto size {0};
+        std::streamoff size {0};
         std::unique_ptr<char[]> spBuffer;
         std::ifstream file {filePath, std::ios_base::binary};
 
         if (file.is_open())
         {
             // Get pointer to associated buffer object
-            auto buffer {file.rdbuf()};
+            auto* buffer = file.rdbuf();
 
             if (nullptr != buffer)
             {
                 // Get file size using buffer's members
                 size = buffer->pubseekoff(0, file.end, file.in);
+                if (size < 0)
+                {
+                    return std::vector<char> {};
+                }
                 buffer->pubseekpos(0, file.in);
                 // Allocate memory to contain file data
-                spBuffer = std::make_unique<char[]>(size);
+                auto size_t_size = static_cast<std::size_t>(size);
+                spBuffer = std::make_unique<char[]>(size_t_size);
                 // Get file data
-                buffer->sgetn(spBuffer.get(), size);
+                buffer->sgetn(spBuffer.get(), static_cast<std::streamsize>(size_t_size));
             }
         }
 
-        return std::vector<char> {spBuffer.get(), spBuffer.get() + size};
+        return std::vector<char> {spBuffer.get(), spBuffer.get() + static_cast<std::size_t>(size)};
     }
 }; // namespace file_io
